Fixed null pointer error in deleteOrganization and missing test stubs for delete-need/delete-org

+++ ufund-api/src/main/java/com/ufund/api/ufundapi/controller/OrganizationController.java
@@ -142,8 +142,11 @@ public class OrganizationController {
                 return new ResponseEntity<>(HttpStatus.FORBIDDEN);
             }
 
-            for (int needId : oService.getOrganization(name).getNeeds().keySet()) {
-                needService.deleteNeed(needId);
+            Organization org = oService.getOrganization(name);
+            if (org != null) {
+                for (int needId : org.getNeeds().keySet()) {
+                    needService.deleteNeed(needId);
+                }
             }
 
             boolean updated = oService.deleteOrganization(name);
